refactor error tag component, hook, utils
