+++ src/components/ErrorTagComment/ErrorTagComment.js
@@ -1,0 +1,20 @@
+import useErrorTagOptions from '../../hooks/UseErrorTagOptions';
+import { formatErrorTags } from '../../utils/errorTagUtils';
+
+const ErrorTagComment = ({ errorTags }) => {
+  const options = useErrorTagOptions();
+
+  if (options.data) {
+    const formattedErrorTags = formatErrorTags(errorTags, options);
+  
+    if (formattedErrorTags) {
+      const str = formattedErrorTags.length > 1 ? formattedErrorTags.join(", ") : formatErrorTags;
+
+      return str;
+    }
+  }
+
+  return null;
+}
+
+export default ErrorTagComment
